+++ scripts/train_gemma.py
@@ -1,8 +1,7 @@
 """
 tested with:
+
 pip install unsloth==2025.5.7 unsloth-zoo==2025.5.8
-
-
 """
 
 from opensloth.opensloth_config import (
@@ -19,32 +18,12 @@ GLOBAL_BZ = 32
 
 DEVICES = [0, 2]
 
-BZ = 2  # if sequence packing, then should be 1, larger does not contribute to speed
+BZ = 2
+# You need to first cache the dataset:
+# RUN: python cache_unsloth_dataset/gemma.py
+
 opensloth_config = OpenSlothConfig(
-    # Use Hugging Face dataset configuration
-    data=HFDatasetConfig(
-        tokenizer_name="unsloth/gemma-3-4b-it-unsloth-bnb-4bit",
-        chat_template="gemma3",  # known template: ['unsloth', 'zephyr', 'chatml', 'mistral', 'llama', 'vicuna', 'vicuna_old', 'vicuna old', 'alpaca', 'gemma', 'gemma_chatml', 'gemma2', 'gemma2_chatml', 'llama-3', 'llama3', 'phi-3', 'phi-35', 'phi-3.5', 'llama-3.1', 'llama-31', 'llama-3.2', 'llama-3.3', 'llama-32', 'llama-33', 'qwen-2.5', 'qwen-25', 'qwen25', 'qwen2.5', 'phi-4', 'gemma-3', 'gemma3', 'qwen-3', 'qwen3']
-        instruction_part="<start_of_turn>user\n",
-        response_part="<start_of_turn>model\n",
-        num_samples=100,
-        nproc=1,
-        max_seq_length=16000,
-        source_type="hf",
-        dataset_name="mlabonne/FineTome-100k",
-        split="train",
-        cache=True,
-    ),
-    # Use PathDatasetConfig if you have a local dataset (only support sharegpt format for now)
-    # data=PathDatasetConfig(
-    #     path="sharegpt-format-dataset.json",  # we've just saved
-    #     chat_template="qwen3",
-    #     instruction_part="<|im_start|>user\n",
-    #     response_part="<|im_start|>assistant\n",
-    #     num_samples=10000,
-    #     nproc=52,
-    #     max_seq_length=16000,
-    # ),
+    data=HFDatasetConfig(cache_path="data/cache_gemma_responses_only"),
     devices=DEVICES,
     fast_model_args=FastModelArgs(
         model_name="unsloth/gemma-3-4b-it-unsloth-bnb-4bit",
@@ -84,6 +63,7 @@ training_config = TrainingArguments(
     optim="adamw_8bit",
     seed=3407,
     report_to="none",  # or wandb/tensorboard
+    max_steps=100,
 )